Actualizacion 2 services

+++ src/main/java/model/services/AerolineaServices.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface AerolineaServices {
     AerolineaDTO save(AerolineaDTO airline);
-    Optional<AerolineaDTO > findById(int id);
-    Optional<AerolineaDTO > update(int id, AerolineaDTO  airline);
+    Optional<AerolineaDTO > findById(Long id);
+    Optional<AerolineaDTO > update(Long id, AerolineaDTO  airline);
     List<AerolineaDTO > findAll();
     List<AerolineaDTO > findByName(String name);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
